roll-back dot-syntax to its own module and add test for syntax
