+++ core/shared/src/main/scala/monocle/Fold.scala
@@ -1,7 +1,5 @@
 package monocle
 
-import monocle.function._
-
 trait Fold[A, B] { self =>
   def toIterator(from: A): Iterator[B]
 
@@ -61,40 +59,6 @@ trait Fold[A, B] { self =>
 
   def asTarget[C](implicit ev: B =:= C): Fold[A, C] =
     asInstanceOf[Fold[A, C]]
-
-  ///////////////////////////////////
-  // dot syntax for optics typeclass
-  ///////////////////////////////////
-
-  def _1(implicit ev: Field1[B]): Fold[A, ev.B] = first(ev)
-  def _2(implicit ev: Field2[B]): Fold[A, ev.B] = second(ev)
-  def _3(implicit ev: Field3[B]): Fold[A, ev.B] = third(ev)
-  def _4(implicit ev: Field4[B]): Fold[A, ev.B] = fourth(ev)
-  def _5(implicit ev: Field5[B]): Fold[A, ev.B] = fifth(ev)
-  def _6(implicit ev: Field6[B]): Fold[A, ev.B] = sixth(ev)
-
-  def first(implicit ev: Field1[B]): Fold[A, ev.B]  = compose(ev.first)
-  def second(implicit ev: Field2[B]): Fold[A, ev.B] = compose(ev.second)
-  def third(implicit ev: Field3[B]): Fold[A, ev.B]  = compose(ev.third)
-  def fourth(implicit ev: Field4[B]): Fold[A, ev.B] = compose(ev.fourth)
-  def fifth(implicit ev: Field5[B]): Fold[A, ev.B]  = compose(ev.fifth)
-  def sixth(implicit ev: Field6[B]): Fold[A, ev.B]  = compose(ev.sixth)
-
-  def at[I, C](i: I)(implicit ev: At.Aux[B, I, C]): Fold[A, Option[C]] = compose(ev.at(i))
-  def cons(implicit ev: Cons[B]): Fold[A, (ev.B, B)]                   = compose(ev.cons)
-  def headOption(implicit ev: Cons[B]): Fold[A, ev.B]                  = compose(ev.headOption)
-  def tailOption(implicit ev: Cons[B]): Fold[A, B]                     = compose(ev.tailOption)
-  def index[I, C](i: I)(implicit ev: Index.Aux[B, I, C]): Fold[A, C]   = compose(ev.index(i))
-  def possible(implicit ev: Possible[B]): Fold[A, ev.B]                = compose(ev.possible)
-  def reverse(implicit ev: Reverse[B]): Fold[A, ev.B]                  = compose(ev.reverse)
-
-  ///////////////////////////////////
-  // dot syntax for standard types
-  ///////////////////////////////////
-
-  def left[E, C](implicit ev: B =:= Either[E, C]): Fold[A, E]  = asTarget[Either[E, C]].compose(Prism.left[E, C])
-  def right[E, C](implicit ev: B =:= Either[E, C]): Fold[A, C] = asTarget[Either[E, C]].compose(Prism.right[E, C])
-  def some[C](implicit ev: B =:= Option[C]): Fold[A, C]        = asTarget[Option[C]].compose(Prism.some[C])
 }
 
 object Fold {
