+++ core/shared/src/main/scala/monocle/Lens.scala
@@ -1,6 +1,6 @@
 package monocle
 
-import monocle.function._
+import monocle.function.{At, Field1, Field2, Field3, Field4, Field5, Field6}
 
 trait Lens[A, B] extends Optional[A, B] with Getter[A, B] { self =>
 
@@ -16,37 +16,6 @@ trait Lens[A, B] extends Optional[A, B] with Getter[A, B] { self =>
     def set(to: C): A => A                 = self.modify(other.set(to))
     override def modify(f: C => C): A => A = self.modify(other.modify(f))
   }
-
-  ///////////////////////////////////
-  // dot syntax for optics typeclass
-  ///////////////////////////////////
-
-  override def _1(implicit ev: Field1[B]): Lens[A, ev.B] = first(ev)
-  override def _2(implicit ev: Field2[B]): Lens[A, ev.B] = second(ev)
-  override def _3(implicit ev: Field3[B]): Lens[A, ev.B] = third(ev)
-  override def _4(implicit ev: Field4[B]): Lens[A, ev.B] = fourth(ev)
-  override def _5(implicit ev: Field5[B]): Lens[A, ev.B] = fifth(ev)
-  override def _6(implicit ev: Field6[B]): Lens[A, ev.B] = sixth(ev)
-
-  override def first(implicit ev: Field1[B]): Lens[A, ev.B]  = compose(ev.first)
-  override def second(implicit ev: Field2[B]): Lens[A, ev.B] = compose(ev.second)
-  override def third(implicit ev: Field3[B]): Lens[A, ev.B]  = compose(ev.third)
-  override def fourth(implicit ev: Field4[B]): Lens[A, ev.B] = compose(ev.fourth)
-  override def fifth(implicit ev: Field5[B]): Lens[A, ev.B]  = compose(ev.fifth)
-  override def sixth(implicit ev: Field6[B]): Lens[A, ev.B]  = compose(ev.sixth)
-
-  override def at[I, C](i: I)(implicit ev: At.Aux[B, I, C]): Lens[A, Option[C]] = compose(ev.at(i))
-  override def reverse(implicit ev: Reverse[B]): Lens[A, ev.B]                  = compose(ev.reverse)
-
-  ///////////////////////////////////
-  // dot syntax for standard types
-  ///////////////////////////////////
-
-  override def left[E, C](implicit ev: B =:= Either[E, C]): Optional[A, E] =
-    asTarget[Either[E, C]].compose(Prism.left[E, C])
-  override def right[E, C](implicit ev: B =:= Either[E, C]): Optional[A, C] =
-    asTarget[Either[E, C]].compose(Prism.right[E, C])
-  override def some[C](implicit ev: B =:= Option[C]): Optional[A, C] = asTarget[Option[C]].compose(Prism.some[C])
 }
 
 object Lens {
