--- core/shared/src/main/scala/monocle/implicits.scala
@@ -1,3 +1,0 @@
-package monocle
-
-object implicits extends syntax.AppliedSyntax
