+++ dotSyntax/src/main/scala/monocle/syntax/all.scala
@@ -1,0 +1,11 @@
+package monocle.syntax
+
+object all
+    extends AppliedSyntax
+    with FoldSyntax
+    with GetterSyntax
+    with IsoSyntax
+    with LensSyntax
+    with OptionalSyntax
+    with PrismSyntax
+    with SetterSyntax
